Updating checkstyle and PMD versions. Refactor code
