+++ .travis.yml
@@ -7,9 +7,9 @@ branches:
 only: develop
 
 before_script:
-  - wget http://services.gradle.org/distributions/gradle-4.0-bin.zip
-  - unzip gradle-4.0-bin.zip
-  - export GRADLE_HOME=$PWD/gradle-4.0
+  - wget http://services.gradle.org/distributions/gradle-4.1-bin.zip
+  - unzip gradle-4.1-bin.zip
+  - export GRADLE_HOME=$PWD/gradle-4.1
   - export PATH=$GRADLE_HOME/bin:$PATH
 
 script: gradle check
