+++ src/main/java/org/fundacionjala/coding/marcos/AveragesOfNumbers.java
@@ -1,4 +1,4 @@
-package org.fundacionjala.coding.marcoslara;
+package org.fundacionjala.coding.marcos;
 
 /**
  * Created by Marcos Lara on 6/6/2017.
