+++ src/main/java/org/fundacionjala/coding/marcos/EanValidator.java
@@ -1,4 +1,4 @@
-package org.fundacionjala.coding.marcoslara;
+package org.fundacionjala.coding.marcos;
 
 /**
  * Created by Marcos Lara on 5/26/2017.
