+++ src/main/java/org/fundacionjala/coding/marcos/HighestAndLowest.java
@@ -1,4 +1,4 @@
-package org.fundacionjala.coding.marcoslara;
+package org.fundacionjala.coding.marcos;
 
 import java.util.Arrays;
 import java.util.Collections;
