+++ src/main/java/org/fundacionjala/coding/marcos/MultiplesOfThreeAndFive.java
@@ -1,4 +1,4 @@
-package org.fundacionjala.coding.marcoslara;
+package org.fundacionjala.coding.marcos;
 
 /**
  * Created by Marcos on 12-06-2017.
