+++ src/main/java/org/fundacionjala/coding/marcos/Sequence.java
@@ -1,4 +1,4 @@
-package org.fundacionjala.coding.marcoslara;
+package org.fundacionjala.coding.marcos;
 
 /**
  * Created by Marcos on 18-08-2017.
