+++ src/main/java/org/fundacionjala/coding/marcos/SpinWords.java
@@ -1,4 +1,4 @@
-package org.fundacionjala.coding.marcoslara;
+package org.fundacionjala.coding.marcos;
 
 /**
  * Created by Marcos Lara on 5/22/2017.
