+++ src/test/java/org/fundacionjala/coding/marcos/SequenceTest.java
@@ -1,4 +1,4 @@
-package org.fundacionjala.coding.marcoslara;
+package org.fundacionjala.coding.marcos;
 
 import java.lang.reflect.Constructor;
 import java.lang.reflect.InvocationTargetException;
